Combine AnnotatioWrapper and AnnotatioProvider
