+++ packages/core/src/components/context/AnnotationProvider.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import { NewExamAnnotation } from '../../types/Score'
 import { AnnotationProps } from '../exam/Exam'
+import { onMouseDownForAnnotation } from '../grading/examAnnotationUtils'
+import { AnnotationPopup } from '../shared/AnnotationPopup'
 
 interface Props {
   children: React.ReactNode
@@ -24,6 +26,21 @@ export const AnnotationProvider = ({
   const [newAnnotation, setNewAnnotation] = React.useState<NewExamAnnotation | null>(null)
   const [newAnnotationRef, setNewAnnotationRef] = React.useState<HTMLElement>()
 
+  const mouseUpCallback = (annotation: NewExamAnnotation) => {
+    console.log('UP')
+    setNewAnnotation(annotation)
+  }
+
+  function onMouseDown(e: React.MouseEvent) {
+    console.log('DOWN')
+    const target = e.target as Element
+    const clickIsInPopup = target.closest('.annotation-popup')
+    if (!clickIsInPopup) {
+      setNewAnnotation(null)
+    }
+    onMouseDownForAnnotation(e, mouseUpCallback)
+  }
+
   if (!onClickAnnotation && !onSaveAnnotation) {
     return children
   }
@@ -34,13 +51,16 @@ export const AnnotationProvider = ({
         annotations,
         onClickAnnotation,
         onSaveAnnotation,
-        setNewAnnotation,
         newAnnotation,
+        setNewAnnotation,
         newAnnotationRef,
         setNewAnnotationRef
       }}
     >
-      {children}
+      <span onMouseDown={onMouseDown}>
+        <AnnotationPopup />
+        {children}
+      </span>
     </AnnotationContext.Provider>
   )
 }
