+++ packages/core/src/components/grading-instructions/GradingInstructions.tsx
@@ -1,37 +1,36 @@
 import React, { useContext, useEffect } from 'react'
-import { AnnotationProps, AnnotationWrapper, CommonExamProps } from '../exam/Exam'
-import { CommonExamContext, withCommonExamContext } from '../context/CommonExamContext'
-import { useCached } from '../../useCached'
-import { changeLanguage, initI18n, useExamTranslation } from '../../i18n'
-import { scrollToHash } from '../../scrollToHash'
 import { I18nextProvider } from 'react-i18next'
-import Section from './Section'
-import SectionTitle from './SectionTitle'
-import Question from './Question'
-import File from '../shared/File'
-import ExamGradingInstruction from './ExamGradingInstruction'
-import AutogradedAnswer from './AutogradedAnswer'
-import Image from '../shared/Image'
-import Formula from '../shared/Formula'
-import AutogradedAnswerOption from './AutogradedAnswerOption'
-import QuestionTitle from './QuestionTitle'
 import { createRenderChildNodes } from '../../createRenderChildNodes'
-import { examTitleId } from '../../ids'
-import DocumentTitle from '../DocumentTitle'
-import SectionElement from '../SectionElement'
 import { findChildElement, queryAncestors } from '../../dom-utils'
-import { renderIf } from '../RenderIf'
+import { changeLanguage, initI18n, useExamTranslation } from '../../i18n'
+import { examTitleId } from '../../ids'
+import { scrollToHash } from '../../scrollToHash'
+import { useCached } from '../../useCached'
+import DocumentTitle from '../DocumentTitle'
 import RenderExamElements from '../RenderExamElements'
-import { mkTableOfContents } from '../shared/TableOfContents'
-import AnswerGradingInstruction from './AnswerGradingInstruction'
+import { renderIf } from '../RenderIf'
+import SectionElement from '../SectionElement'
+import { AnnotationProvider } from '../context/AnnotationProvider'
+import { CommonExamContext, withCommonExamContext } from '../context/CommonExamContext'
+import { AnnotationProps, CommonExamProps } from '../exam/Exam'
+import QuestionInstruction from '../exam/QuestionInstruction'
 import mkAttachmentLink from '../shared/AttachmentLink'
 import mkAttachmentLinks from '../shared/AttachmentLinks'
-import Recording from './Recording'
-import QuestionInstruction from '../exam/QuestionInstruction'
-import { QuestionNumber } from '../shared/QuestionNumber'
 import ExamTranslation from '../shared/ExamTranslation'
-import { AnnotationProvider } from '../context/AnnotationProvider'
-import { AnnotationPopup } from '../shared/AnnotationPopup'
+import File from '../shared/File'
+import Formula from '../shared/Formula'
+import Image from '../shared/Image'
+import { QuestionNumber } from '../shared/QuestionNumber'
+import { mkTableOfContents } from '../shared/TableOfContents'
+import AnswerGradingInstruction from './AnswerGradingInstruction'
+import AutogradedAnswer from './AutogradedAnswer'
+import AutogradedAnswerOption from './AutogradedAnswerOption'
+import ExamGradingInstruction from './ExamGradingInstruction'
+import Question from './Question'
+import QuestionTitle from './QuestionTitle'
+import Recording from './Recording'
+import Section from './Section'
+import SectionTitle from './SectionTitle'
 
 const renderIfWithinGradingInstructionContent = renderIf(
   ({ element }) =>
@@ -112,7 +111,6 @@ const GradingInstructions: React.FunctionComponent<CommonExamProps & AnnotationP
     >
       <I18nextProvider i18n={i18n}>
         <main className="e-exam e-grading-instructions" lang={subjectLanguage}>
-          <AnnotationPopup />
           <React.StrictMode />
           <SectionElement aria-labelledby={examTitleId}>
             <DocumentTitle id={examTitleId}>
@@ -124,9 +122,7 @@ const GradingInstructions: React.FunctionComponent<CommonExamProps & AnnotationP
               </p>
             )}
             {examGradingInstruction && (
-              <AnnotationWrapper>
-                <ExamGradingInstruction {...{ element: examGradingInstruction, renderChildNodes }} />
-              </AnnotationWrapper>
+              <ExamGradingInstruction {...{ element: examGradingInstruction, renderChildNodes }} />
             )}
             {tableOfContents && (
               <div className="main-toc-container">
@@ -134,7 +130,7 @@ const GradingInstructions: React.FunctionComponent<CommonExamProps & AnnotationP
               </div>
             )}
           </SectionElement>
-          <AnnotationWrapper>{renderChildNodes(root)}</AnnotationWrapper>
+            {renderChildNodes(root)}
         </main>
       </I18nextProvider>
     </AnnotationProvider>
