+++ packages/core/src/components/grading/examAnnotationUtils.ts
@@ -2,21 +2,18 @@ import React from 'react'
 import { textAnnotationFromRange } from './editAnnotations'
 
 export function onMouseDownForAnnotation(e: React.MouseEvent, mouseUpCallback: (e: any) => void) {
-  function onMouseUpAfterAnswerMouseDown(e: MouseEvent) {
-    const target = e.target as HTMLElement
-    console.log(target)
+  function onMouseUpAfterAnswerMouseDown() {
     window.removeEventListener('mouseup', onMouseUpAfterAnswerMouseDown)
 
-    // Text annotation
+    /**
+     * selection.anchorNode = where mouse is pressed down (chronologically first)
+     * selection.getRangeAt(0).startContainer = where selection starts (first in DOM)
+     */
+
     const selection = window.getSelection()
     const startNode = selection?.anchorNode?.parentElement
     const endNode = selection?.focusNode?.parentElement
-    const startNodePath = selection?.anchorNode?.parentElement?.getAttribute('data-annotation-path')
-    const endNodePath = selection?.focusNode?.parentElement?.getAttribute('data-annotation-path')
-    const range2 = selection?.getRangeAt(0)
-    console.log('startnode path', startNodePath)
-    console.log('R', range2)
-    console.log('same?', startNode === endNode && startNode !== null)
+    const endNodePath = endNode?.getAttribute('data-annotation-path')
 
     if (selection && endNodePath && hasOnlyOneTextElementSelected()) {
       const range = selection.getRangeAt(0)
@@ -24,8 +21,6 @@ export function onMouseDownForAnnotation(e: React.MouseEvent, mouseUpCallback: (
         startNode?.parentElement?.closest('div[data-annotation-anchor]')?.getAttribute('data-annotation-anchor') || ''
 
       const length = Math.abs(selection.anchorOffset - selection.focusOffset)
-
-      console.log('Tgt', target, selection?.anchorNode?.parentElement, selection?.focusNode?.parentElement)
 
       const position1 = textAnnotationFromRange(selection.focusNode?.parentElement as HTMLElement, range)
       if (!position1 || !length) {
@@ -51,8 +46,14 @@ export function onMouseDownForAnnotation(e: React.MouseEvent, mouseUpCallback: (
 
 function hasOnlyOneTextElementSelected(): boolean {
   const selection = window.getSelection()
-  const range = selection?.getRangeAt(0)
-  return range?.startContainer.parentElement === range?.endContainer.parentElement
+  const visibleMarkTagExistsInSelection =
+    selection &&
+    Array.from(selection?.getRangeAt(0).cloneContents().children).some(
+      child => child.tagName === 'MARK' && child.getAttribute('data-hidden') === 'false'
+    )
+  const startNode = selection?.anchorNode?.parentElement
+  const endNode = selection?.focusNode?.parentElement
+  return !visibleMarkTagExistsInSelection && startNode === endNode
 }
 
 export function hasTextSelectedInAnswerText(): boolean {
